Re-organised files & folders, separated out the data objects from CombatParser.py
