+++ Source/combat/CombatParser.py
@@ -2,7 +2,10 @@ from re import compile
 import os.path
 import time
 from datetime import datetime
-import random
+from data.Ability import Ability
+from data.Character import Character
+from data.DamageComponent import DamageComponent
+from combat.CombatSession import CombatSession
 from tkinter import filedialog as tk
 import sys
 from PyQt5.QtCore import QObject, pyqtSignal, pyqtSlot, QMutex, QMutexLocker, QTimer, QCoreApplication
@@ -315,7 +318,7 @@ class Parser(QObject):
             if Globals.CONSOLE_VERBOSITY >= 3:
                 print(f"     Added Ability: {search_ability} to Character: {char.get_name()} via Hit Roll Event")
 
-        # if pet: char.get_ability(search_ability).ability_used() # We don't have power activation events for pets, so we'll use the hit roll as a proxy for ability usage
+        if pet: char.get_ability(search_ability).ability_used() # We don't have power activation events for pets, so we'll use the hit roll as a proxy for ability usage
         char.get_ability(search_ability).ability_hit(data["outcome"] == "HIT") 
     def handle_event_player_damage(self, data, pet=False):
         '''Handles a player damage event'''
@@ -490,7 +493,6 @@ class Parser(QObject):
                     refresher = 0
                     QCoreApplication.processEvents()
                     if not self.processing_live: break
-        if CLI_MODE: _test_show_results()
         print('          Log File processed in: ', round(time.time() - _log_process_start_, 2), ' seconds')
         self.sig_finished.emit(self.combat_session_data)
         return True
@@ -608,463 +610,3 @@ class Parser(QObject):
         user_session_name = ""
 
         if Globals.CONSOLE_VERBOSITY >= 2: print('          Parser variables cleaned...')
-    
-class CombatSession(QObject):
-    '''The CombatSession class stores data about a combat session, which is a period where damage events are registered.
-    CombatSessions will automatically end based on the COMBAT_SESSION_TIMEOUT value to avoid including long downtime periods in the data.''' 
-    def __init__(self, timestamp=0, name=""):
-        super().__init__()
-        self.start_time = timestamp
-        self.end_time = timestamp
-        self.duration = 0 # Seconds
-        self.chars = {} # Stores a list of characters
-        self.exp_value = 0
-        self.inf_value = 0
-        self.name = name
-        
-
-    def set_start_time(self, start_time):
-        self.start_time = start_time
-    def set_end_time(self, end_time):
-        self.end_time = end_time
-    def set_name(self, name):
-        self.name = name
-    def update_duration(self):
-        self.duration = self.end_time - self.start_time
-    def update_session_time(self, timestamp):
-        self.end_time = timestamp
-        self.update_duration()
-
-    def get_name(self):
-        return self.name
-    def get_duration(self):
-        self.update_duration()
-        return self.duration
-    
-    def get_total_damage(self):
-        '''Calculates the total damage for the session'''
-        sum = 0
-        for char in self.chars:
-            sum += self.chars[char].get_total_damage()
-        return round(sum,2)
-    
-    def get_average_damage(self):
-        '''Calculates the average damage for the session'''
-        damage = self.get_total_damage()
-        if damage > 0 : damage =  round(damage / self.get_count(),2)
-        return damage
-    
-    def get_dps(self):
-        '''Calculates the DPS for the session based off of the damage components'''
-        dps = 0
-        for char in self.chars:
-            dps += self.chars[char].get_dps(self.duration)
-        return round(dps,2)
-    
-    def add_exp(self, exp_value):
-        '''Adds the exp value to the total exp'''
-        self.exp_value += exp_value
-    def add_inf(self, inf_value):
-        '''Adds the inf value to the total inf'''
-        self.inf_value += inf_value
-    def get_exp(self):
-        '''Returns the total exp'''
-        return self.exp_value
-    def get_inf(self):
-        '''Returns the total inf'''
-        return self.inf_value
-    
-    def add_character(self, character):
-        self.chars[character.get_name()] = character
-
-class Character(QObject):
-    '''Stores data about a character, this can be the Player, pets or enemies'''
-    def __init__(self, name="") -> None:
-        super().__init__()
-        self.name = name
-        self.abilities = {}
-        self.is_pet = False
-
-    def add_ability(self, ability_name, ability):
-        self.abilities[ability_name] = ability
-
-    def get_ability(self, ability_name):
-        if ability_name not in self.abilities:
-            return ""
-        return self.abilities[ability_name]
-    
-    def get_name(self):
-        return self.name
-    
-    def get_total_damage(self):
-        '''Calculates the total damage for the character'''
-        sum = 0
-        if self.abilities == {}: return sum
-        for ability in self.abilities:
-            sum += self.abilities[ability].get_total_damage()
-        return round(sum,2)
-    
-    def get_average_damage(self):
-        '''Calculates the overall average damage per hit for the character by looping through all abilities' average damage and returning the sum'''
-        average = 0
-        ability_count = 0
-        if self.abilities == {}: return average
-        for ability in self.abilities:
-            average += self.abilities[ability].get_average_damage()
-            ability_count += 1
-        if ability_count  > 0 and average > 0:
-            return round(average / ability_count,2)
-        return average
-
-    def get_accuracy(self):
-        '''Calculates the accuracy for the character'''
-        if self.abilities == {}: return 0
-        hits = 0
-        tries = 0
-        for ability in self.abilities:
-            hits += self.abilities[ability].get_hits()
-            tries += self.abilities[ability].get_tries()
-        if tries == 0: return 0
-        return round((hits / tries) * 100,2)
-    
-    def get_dps(self, duration):
-        '''Calculates the DPS for the character based off of the damage components'''
-        dps = 0
-        if self.abilities == {}: return dps
-        for ability in self.abilities:
-            dps += self.abilities[ability].get_dps(duration)
-        return round(dps,2)
-    
-    def get_count(self):
-        '''Returns the number of times the character has been used'''
-        count = 0
-        if self.abilities == {}: return count
-        for ability in self.abilities:
-            count += self.abilities[ability].get_count()
-        return count
-    
-class Ability(QObject):
-    '''Stores data about an ability used'''
-    def __init__(self, name, hit=None, proc=False):
-        super().__init__()
-        self.name = name
-        self.count = 0
-        self.tries = 0 if hit is None else 1
-        self.hits = 1 if hit is True else 0
-        self.damage = []
-        self.proc = proc
-        self.pet = False
-        self.pet_name = "" # If ability came from a pet, store the pet name
-
-    def add_damage(self, damage_component, value = 0):
-        '''Adds a damage component to the ability'''
-        # Check if damage type is already in the ability, if it is, add the damage to the existing component
-        for component in self.damage:
-            if component.type == damage_component.type:
-                component.add_damage(damage_component.type, value)
-                if self.proc: 
-                    self.ability_used()
-                    
-                #print ('Added to existing damage component to ', self.name, ' | ', damage_component.type, ' | ', self.get_total_damage(), ' | ', self.damage[-1].get_count())
-                return
-            
-        self.damage.append(damage_component)
-        self.damage[-1].add_damage(damage_component.type, value)
-        #print ('Added new damage component to ', self.name, ' | ', damage_component.type, ' | ', self.get_total_damage(), ' | ', self.damage[-1].get_count())
-
-        if self.proc:
-            self.ability_used() 
-    
-    def ability_used(self):
-        '''Handles incrementing the count for each time ability is used. Can also be called during a hit check to increment the hit counter'''
-        self.count += 1
-    def ability_hit(self, hit):
-        self.tries += 1
-        if hit: self.hits += 1
-    def get_name(self):
-        '''Returns the name of the ability'''
-        return self.name
-    
-    def get_total_damage(self):
-        '''Calculates the total damage for the ability'''
-        sum = 0
-        for component in self.damage:
-            sum += component.get_damage()
-            #print('Damage Component for ', self.name,': ', component.type,'|', component.total_damage,'|', component.count)
-        return round(sum,2)
-    def get_max_damage(self):
-        '''Returns the highest damage for the ability'''
-        highest = 0
-        for component in self.damage:
-            highest += component.get_highest_damage()
-        return highest
-    def get_min_damage(self):
-        '''Returns the lowest damage for the ability'''
-        lowest = 0
-        for component in self.damage:
-            lowest = component.get_lowest_damage()
-        return lowest
-    def get_average_damage(self):
-        '''Calculates the average damage for the ability'''
-        average = 0
-        for component in self.damage:
-            average += component.get_average_damage()
-        return round(average,2)
-    
-    def get_accuracy(self):
-        '''Calculates the accuracy for the ability'''
-        if self.hits == 0 or self.tries == 0:
-            return 0
-        return round((self.hits / self.tries) * 100,2)
-    
-    def get_count(self):
-        '''Returns the number of times the ability has been used'''
-        return self.count
-    
-    def get_dps(self, duration):
-        '''Calculates the DPS for the ability based off of the damage components'''
-        dps = 0
-        for component in self.damage:
-            dps += component.get_dps(duration)
-        return round(dps,2)
-
-    def get_hits(self):
-        '''Returns the number of times the ability has hit'''
-        return self.hits
-    def get_tries(self):
-        '''Returns the number of times the ability has been used'''
-        return self.tries
-class DamageComponent(QObject):
-    '''Stores data about a damage component'''
-    def __init__(self, type="", value : float = 0):
-        super().__init__()
-        self.count = 0
-        self.type = type
-        self.name = type
-        self.total_damage = value
-        self.highest_damage = 0
-        self.lowest_damage = 0
-    
-    def add_damage(self, type, value : float):
-        '''Adds damage value to the damage component'''
-        self.count += 1
-        self.total_damage += value
-        self.total_damage = round(self.total_damage, 2)
-        self.update_min_max_damage(value)
-    
-    def get_dps(self, duration):
-        '''Calculates the DPS for the damage component'''
-        if duration == 0: return self.total_damage
-        else: return round(self.total_damage / duration, 2)
-    def get_average_damage(self):
-        '''Calculates the average damage for the damage component'''
-        if self.count == 0: return self.total_damage
-        return round(self.total_damage / self.count,2)
-    def get_highest_damage(self):
-        '''Returns the highest damage for the damage component'''
-        return self.highest_damage
-    def get_lowest_damage(self):
-        '''Returns the lowest damage for the damage component'''
-        return self.lowest_damage
-    def get_damage(self):
-        '''Returns the total damage for the damage component'''
-        return self.total_damage
-    def get_damage_component(self):
-        return self
-    def get_count(self):
-        '''Returns the number of times the damage component has been used'''
-        return self.count
-    def update_min_max_damage(self, value : float):
-        '''Updates the min and max damage for the damage component'''
-        if value >= self.highest_damage or self.highest_damage == 0:
-            self.highest_damage = value
-            #print ('         Highest Damage Updated: ', self.highest_damage)
-        if value <= self.lowest_damage or self.lowest_damage == 0:
-            self.lowest_damage = value
-            #print ('         Lowest Damage Updated: ', self.highest_damage)
-    
-class Target:
-    '''Stores information and data about a target'''        
-    def __init__(self, name):
-        pass
-
-
-def _test_hit_rolls():
-        sample_hit_lines = [
-    '2023-11-18 14:49:49 MISSED Ugly Face Mother!! Your Inky Aspect power had a 95.00% chance to hit, you rolled a 95.27.\n',
-    '2023-11-18 14:49:51 HIT Thorn! Your Void Total Radial Judgement power had a 95.00% chance to hit, you rolled a 41.38.\n',
-    '2023-11-18 14:49:51 HIT Lattice! Your Void Total Radial Judgement power had a 95.00% chance to hit, you rolled a 59.66.\n',
-    '2023-11-18 14:49:53 HIT Your Worst Nightmare! Your Black Dwarf Mire power had a 95.00% chance to hit, you rolled a 68.99.\n',
-    '2023-11-18 14:49:53 HIT Lattice! Your Black Dwarf Mire power had a 95.00% chance to hit, you rolled a 54.51.\n',
-    '2023-11-18 14:49:54 MISSED Thug!! Your Inky Aspect power had a 95.00% chance to hit, you rolled a 99.40.\n',
-    '2023-11-18 14:50:07 HIT Mighty Golem! Your Dark Detonation power was forced to hit by streakbreaker.\n',
-    '2023-11-18 14:49:55 HIT Lattice! Your Dark Nova Blast power was forced to hit by streakbreaker.\n',
-    '2023-11-18 14:49:56 HIT Lattice! Your Dark Nova Emanation power was forced to hit by streakbreaker.\n',
-    '2023-11-18 14:50:00 MISSED Lattice!! Your Dark Nova Bolt power had a 95.00% chance to hit, you rolled a 95.37.\n',
-    '2023-11-18 14:50:07 MISSED Dark Servant!! Your Dark Detonation power had a 95.00% chance to hit, you rolled a 95.99.'
-
-]   
-    
-        print('\n \n', "---- STARTING HIT ROLL SAMPLE DATA TEST -----", '\n \n')
-        #test extraction and interpretation of hit roll data
-        for line in sample_hit_lines:
-            event, data = self.extract_from_line(line)
-            #self.line_counter += 1
-            if event != "":
-                print(event, data)
-                self.interpret_event(event, data)
-def _test_damage_lines():
-        sample_damage_lines = [
-        "2023-11-18 14:50:23 You hit Lattice with your Black Dwarf Smite for 211.94 points of Smashing damage.",
-        "2023-11-18 14:50:23 You hit Lattice with your Black Dwarf Smite for 378.47 points of Negative Energy damage.",
-        "2023-11-18 14:50:23 You hit D0xx3r L0rD with your Gladiator's Strike: Chance for Smashing Damage for 246.66 points of Smashing damage.",
-        "2023-11-18 14:50:18 You hit Emet Selch with your Inky Aspect for 4.72 points of unresistable Special damage.",
-        "2023-11-18 14:50:19 You hit Lattice with your Bombardment: Chance for Fire Damage for 207.12 points of Fire damage.",
-        "2023-11-18 14:50:19 You hit Big Birb with your Dark Nova Detonation for 352.19 points of Negative Energy damage.",
-        "2023-11-18 14:50:19 You hit Mighty Golem with your Doublehit for 59.35 points of Energy damage.",
-        "2023-11-18 14:50:20 You hit Lattice with your Dark Nova Emanation for 517.01 points of Negative Energy damage.",
-        "2023-11-18 14:50:20 You hit Lattice with your Positron's Blast: Chance for Energy Damage for 142.71 points of Energy damage.",
-        "2023-11-18 14:50:21 You hit Mega MInd with your Doublehit for 238.07 points of Energy damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Black Dwarf Smite for 211.94 points of Smashing damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Black Dwarf Smite for 378.47 points of Negative Energy damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Gladiator's Strike: Chance for Smashing Damage for 246.66 points of Smashing damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Mako's Bite: Chance for Lethal Damage for 176.18 points of Lethal damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Touch of Death: Negative Energy Damage for 140.94 points of Negative Energy damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Ice Mistral's Torment: Chance for Cold Damage for 176.18 points of Cold damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Doublehit for 111.07 points of Energy damage.",
-        "2023-11-18 14:50:23 You hit Lattice with your Dark Detonation for 186.33 points of Negative Energy damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Dark Detonation for 266.9 points of Negative Energy damage.",
-        "2023-11-18 14:50:23 You hit Thorn with your Degenerative Interface for 10.88 points of Toxic damage over time.",
-    ]
-
-
-        print('\n \n', "---- STARTING DAMAGE SAMPLE DATA TEST -----", '\n \n')
-        #test extraction and interpretation of hit roll data
-        for line in sample_damage_lines:
-            event, data = self.extract_from_line(line)
-            #self.line_counter += 1
-            if event != "":
-                print(event, data)
-                self.interpret_event(event, data)
-def _test_reward_lines():
-    sample_reward_lines = [
-    '2023-11-18 14:48:05 You gain 1,525 experience and 763 influence.',
-    '2023-11-18 14:48:15 You gain 2,922 experience and 1,461 influence.',
-    '2023-11-18 14:49:16 You gain 67 experience and 47 influence.',
-    '2023-11-18 14:49:10 You gain 2 experience.',
-    '2023-11-18 14:50:16 You gain 47 influence.',
-    '2023-11-18 14:50:16 You gain 148 influence',
-    '2023-11-19 14:50:16 You gain 2,445 experience.',
-    '2023-11-19 14:52:11 You gain 124,212,130 experience.',
-    '2023-11-19 14:52:11 You gain 136,288,199 influence.',
-    '2023-11-19 14:52:11 You gain 124,212,130 experience and 32,322,103 influence.'
-    ]
-    print('\n \n', "---- STARTING REWARD SAMPLE DATA TEST -----", '\n \n')
-
-    for line in sample_reward_lines:
-        event, data = self.extract_from_line(line)
-        #self.line_counter += 1
-        if event != "":
-            print(event, data)
-            self.interpret_event(event, data)
-def _test_log_file():
-    print('\n \n', "---- STARTING LOG FILE DATA TEST -----", '\n \n')
-#Open text file with test data, similarly test extraction data
-    with open('C:\\Users\\mdore\\OneDrive\\02_Documents\\02_SCRIPTS\\CoH_Log_Parser\\coh_logfile_snippet.txt', 'r') as file:
-        for line in file:
-            event, data = self.extract_from_line(line)
-            self.line_count += 1
-            if event != "":
-                print(event, data)
-                self.interpret_event(event, data)
-
-def _test_show_results():
-    def format_duration(seconds):
-        hours, remainder = divmod(seconds, 3600)
-        minutes, seconds = divmod(remainder, 60)
-        return f'{int(hours)} hours, {int(minutes)} minutes, {int(seconds)} seconds'
-    
-    print('\n','\n',"---- SUMMARY -----")
-    print('Player Name: ', self.PLAYER_NAME)
-    print(f'Total Experience: {self.get_exp():,}')
-    print(f'Total Influence: {self.get_inf():,}')
-    print("---- ABILITY DATA -----")
-    # loop down the ability dictionary and print the ability name, and '(proc)' if the ability is a proc, accuracy % and total damage
-    for ability in self.abilities:
-        if self.abilities[ability].get_count() == 0: continue # skip abilities with no count
-        print(self.abilities[ability].name, ' (proc)' if self.abilities[ability].proc else '',
-              '\n    DPS: ', self.abilities[ability].get_dps(self.global_combat_duration), 'Avg: ', self.abilities[ability].get_average_damage(),
-              '\n    Accuracy: ', self.abilities[ability].get_accuracy(), '% | Count: ', self.abilities[ability].count, '| Hits: ', self.abilities[ability].hits, '|')
-        
-        
-        # loop down the damage components and print the damage type, total damage and count
-        for component in self.abilities[ability].damage:
-            print('        Damage Type: ', component.type,'| T:', component.total_damage,'| H:', component.highest_damage, '| L:', component.lowest_damage, '| Count: ', component.count)
-        print('------------------')
-
-    total_damage = 0
-    print('')
-    for ability in self.abilities:
-        total_damage += self.abilities[ability].get_total_damage()
-    print('Total Damage: ', round(total_damage,2))
-    print('Total Damage using Combat Sessions: ', round(total_damage,2))
-    print('Total Duration of the log:', format_duration(self.get_log_duration()))
-    print('Total Combat Duration:', format_duration(self.global_combat_duration))
-    print('Total Combat Sessions: ', self.session_count)
-    print('Total Damage Per Second: ', self.get_dps(self.global_combat_duration))
-    print('Lines processed: ', self.line_count)
-
-def _test_combat_sessions():
-    pass
-
-def open_file():
-    root = tk.Tk()
-    root.withdraw()
-    file_path = tk.askopenfilename()
-    return file_path
-
-def handle_commands(self, command):
-    '''When run in CLI mode, text commands entered will be processed'''
-    if command.lower == "t":
-        self.process_existing_log('path_to_test_file.txt')
-
-    elif command.lower() in  ["live", "l"]:
-        file_path = open_file()
-        if file_path != "":
-            self.process_live_log(file_path)
-        # print("Live monitoring is not implemented yet.")
-            
-    elif command.lower() in ["analyze", "a", "analyse"]:
-        file_path = open_file()
-        if file_path != "":
-            self.process_existing_log(file_path)
-    
-    elif command.lower() in ["exit", "quit", "q"]:
-        print("Exiting...")
-        sys.exit()
-    else:
-        print("Unknown command. Please enter a valid command.")    
-if __name__ == "__main__":
-    # if this is being run directly, run the test data
-    print("Parser has been directly called as __main__, running test data")
-
-    #PLAYER_NAME = "Emet Selch"
-    self = Parser()
-    #run application
-
-    
-    self.line_count = 0
-    CLI_MODE = True
-
-    #Initiate tests
-    #_test_hit_rolls()
-    #_test_damage_lines()
-    #_test_reward_lines()
-    #test_file = open_file()
-    #self.process_existing_log(test_file)
-
-
-    while True:
-        user_command = input("Enter a command (e.g., 'run_test_file', 'live_monitor', 'analyze_log_file <path_to_log_file>'): ")
-        handle_commands(self, user_command)
-
